Bump version to 1.0.1
Published gem is built off this, but the current tag (I'll update it)
is based on prev commit (7d8e944) because I apparently fucked up committing this
and didn't realize it (I was a little dazed this morning).

Anyway, going to go ahead and commit it now and update the tag.